Move nested types into own files
