+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/PropertyPages/ProfileCommandNames.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+
+#nullable disable
+
+namespace Microsoft.VisualStudio.ProjectSystem.VS.PropertyPages
+{
+    public static class ProfileCommandNames
+    {
+        public const string Project = "Project";
+        public const string IISExpress = "IISExpress";
+        public const string Executable = "Executable";
+        public const string NoAction = "NoAction";
+    }
+}
